+++ scctool/tasks/texttospeech.py
@@ -8,8 +8,7 @@ import requests
 
 import scctool.settings
 
-module_logger = logging.getLogger(
-    'scctool.settings.texttospeech')  # create logger
+module_logger = logging.getLogger(__name__)
 
 
 class TextToSpeech: