+++ CHANGELOG.md
@@ -1,6 +1,44 @@
 # Changelog
 
 
+## v2.5.2 (2018-11-28)
+
+### New
+
+* A (partial) translation to French is now available thanks to chemsed alias Seireitei.
+
+### Changes
+
+* Countdown will now display 'soon™' or a custom text once the countdown is finished - it can be changed in config.ini.
+
+
+## v2.5.1 (2018-11-18)
+
+### Changes
+
+* Countdown will now display 'soon™' or a custom text once the countdown is finished - it can be changed in config.ini.
+
+
+## v2.5.0 (2018-11-18)
+
+### New
+
+* Added StarCraft2 skin/style for Countdown.
+
+* New Countdown Browser Source!
+
+
+## v2.4.3 (2018-11-11)
+
+### New
+
+* Playernames can now be assigned to Aligulac IDs in the Misc Settings to improve reliability of the Aligulac Prediction.
+
+### Changes
+
+* Improved automatic player selection of Aligulac Prediction.
+
+
 ## v2.4.2 (2018-11-04)
 
 ### Fix